Выполнил задание по организации кода
